Remove unused helper functions
- Remove `_format_for_gdal()` which is never invoked.
- Remove the numba-decorated phase difference helpers `compute_phase_diffs`
- Move the _redirect_unwrapping_log() to dolphin/unwrap/_tophu.py).
- Update type annotations in unwrap._utils
